Add flag to allow for untrimed URL input
This change should, in theory, allow for a users URL input to not be trimmed of trailing slashes.

The upside is that this allows the tool to be useful for monitoring sites with Canonical URLs that include the trailing slash. While it may not be the majority of cases - some servers/sites may be configured to redirect requests to the Canonical URL. So this change should help users in those instances.

The only downside I can think of currently is that a user could now redundantly monitor the same page/site twice. This would be the result if, for example, they configured a url with and without the flag given. However this is something that would be hard to do accidentally - so IMO the risk here is pretty minimal.

Related to #107.